Switch the capability model from MonoCaps to InvisiCaps. This has three effects: it's a 14.5% speed-up on PizBench9004, it allows int-ptr unions to ping-pong between int and ptr, and it makes sizeof(void*)==8. The int-ptr union ping-pong is the most important outcome of this change. It means, for example, that OpenSSH and Lua now require zero changes to work in Fil-C. OpenSSL now only requires build system changes. SQLite only requires a tiny number of changes. The changes in other projects are also significantly reduced. This should make porting code to Fil-C much easier.

+++ filc/tests/zmemmovefail39/zmemmovefail39.c
@@ -1,9 +1,10 @@
 #include <stdfil.h>
+#include <inttypes.h>
 
 int main()
 {
     zptrtable* dst = zptrtable_new();
-    __int128 src;
+    int64_t src;
 
     zmemmove(dst, &src, 16);
